requirements.txt
# Core Dependencies
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine Learning
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
keras==2.13.1

# MLOps
mlflow==2.5.0
dvc==3.13.0
apache-airflow==2.7.0

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.25.0

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2

# Weather Data
cdsapi==0.6.1
netcdf4==1.6.4
xarray==2023.8.0

# Satellite Data (Google Earth Engine)
earthengine-api==0.1.367

# Database
sqlalchemy==1.4.49
pymongo==4.5.0

# AWS
boto3==1.28.25
sagemaker==2.178.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.65.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Docker
gunicorn<21